@@ -4,25 +4,27 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config proxy.conf.json",
-    "build": "ng build",
-    "test": "ng test",
+    "build": "webpack",
+    "test": "karma start ./karma.conf.js",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "protractor ./protractor.conf.js",
+    "start": "webpack-dev-server --port=4200",
+    "pree2e": "webdriver-manager update --standalone false --gecko false --quiet"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.0.0",
-    "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
-    "@angular/material": "^2.0.0-beta.12",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/animations": "^5.0.3",
+    "@angular/cdk": "^5.0.0-rc.1",
+    "@angular/common": "^5.0.3",
+    "@angular/compiler": "^5.0.3",
+    "@angular/core": "^5.0.3",
+    "@angular/forms": "^5.0.3",
+    "@angular/http": "^5.0.3",
+    "@angular/material": "^5.0.0-rc.1",
+    "@angular/platform-browser": "^5.0.3",
+    "@angular/platform-browser-dynamic": "^5.0.3",
+    "@angular/router": "^5.0.3",
+    "@auth0/angular-jwt": "^1.0.0-beta.9",
     "actions-on-google": "^1.2.0",
     "core-js": "^2.4.1",
     "http-status-codes": "^1.3.0",
@@ -31,14 +33,14 @@
     "js-cookie": "^2.2.0",
     "material-design-icons": "^3.0.1",
     "node-fetch": "^1.7.3",
-    "rxjs": "^5.4.1",
+    "rxjs": "^5.5.2",
     "uuid": "^3.1.0",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.9",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/language-service": "^4.0.0",
+    "@angular/cli": "^1.5.2",
+    "@angular/compiler-cli": "^5.0.3",
+    "@angular/language-service": "^5.0.3",
     "@types/express": "^4.0.37",
     "@types/http-status-codes": "^1.0.29",
     "@types/jasmine": "~2.5.53",
@@ -47,7 +49,9 @@
     "@types/node": "~6.0.60",
     "@types/node-fetch": "^1.6.7",
     "@types/uuid": "^3.4.3",
-    "codelyzer": "~3.0.1",
+    "@types/webpack": "^3.8.1",
+    "@types/webpack-dev-middleware": "^1.9.4",
+    "codelyzer": "~4.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -60,6 +64,29 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "2.4.2"
+    "typescript": "2.4.2",
+    "webpack": "~3.8.1",
+    "webpack-dev-middleware": "^1.12.1",
+    "webpack-dev-server": "~2.9.3",
+    "autoprefixer": "^6.5.3",
+    "css-loader": "^0.28.1",
+    "cssnano": "^3.10.0",
+    "exports-loader": "^0.6.3",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^2.29.0",
+    "less-loader": "^4.0.5",
+    "postcss-loader": "^2.0.8",
+    "postcss-url": "^7.1.2",
+    "raw-loader": "^0.5.1",
+    "sass-loader": "^6.0.3",
+    "source-map-loader": "^0.2.0",
+    "istanbul-instrumenter-loader": "^2.0.0",
+    "style-loader": "^0.13.1",
+    "stylus-loader": "^3.0.1",
+    "url-loader": "^0.6.2",
+    "circular-dependency-plugin": "^3.0.0",
+    "webpack-concat-plugin": "1.4.0",
+    "copy-webpack-plugin": "^4.1.1",
+    "uglifyjs-webpack-plugin": "1.0.0"
   }
 }
